working bfs
