+++ bfs.py
@@ -3,94 +3,54 @@ import truck
 import copy
 from moves import Move
 import sys
-import queue
-from collections import deque
 type_of_trash = None
 move_list = []
 s = 0
-q = deque([((), [])])
-# q = queue.Queue()
-move_list = []
-trashes = []
-
-
+sys.setrecursionlimit(12000)
 class BreathFirstSearch:
-
     def start_bfs(self, _map,  _type_of_trash):
         global type_of_trash
         global q
         type_of_trash = _type_of_trash
         starting_pos = ((_map.truck.get_current_position_x(),
                         _map.truck.get_current_position_y()))
-        q = deque([(starting_pos, [])])
-        print("qstart: ", q)
-        self.find_path_bfs(_map, move_list, q)
+        goal = (4, 5)
+        map = copy.deepcopy(_map)
+        self.bfs_shortest_path(map, starting_pos, _map.truck.find_next_trash_to_visit('yellow_trash'))
         return move_list
 
-    def find_path_bfs(self, grid, current_move_list, q):
-        goal = (8, 8)
+    def bfs_shortest_path(self, map, start, goal):
         global move_list
-        # current_q = deque([(start, [])])
-        if self.check_if_is_done(grid):
-            if move_list == [] or len(move_list) > len(current_move_list):
-                print("siemankoo")
-                move_list = current_move_list
-                #current_move_list.clear()
+        if self.check_if_is_done(map):
+            print("done")
             return
-        trash_around = grid.truck.find_trash_around(type_of_trash)
-        if trash_around != []:
-            for i in trash_around:
-                new_grid = copy.deepcopy(grid)
-                new_move_list = copy.deepcopy(current_move_list)
-                new_move_list.append("collect")
-                tmp = new_grid.truck.find_trash_around(type_of_trash)
-                trash_to_collect = tmp[0]
-                new_grid.truck.collect_trash(trash_to_collect)               
-                self.find_path_bfs(new_grid, new_move_list, q)
-        else:
-            curr, path = q.popleft()
-            print("truck pos: ", curr)
-            print("path: ", path)
-            new_grid = copy.deepcopy(grid)
-            new_grid.truck.move_to(curr[0], curr[1])
-            new_move_list = copy.deepcopy(current_move_list)
-            new_move_list.append((curr[0], curr[1]))  
-            possible_cord = new_grid.truck.possible_cord()
-            print("possible:", possible_cord)
-            if possible_cord == []:
-                return
-            for i in possible_cord:
-                if i not in current_move_list and i[0] != 0:
-                    print("move:", i)
-                    print("visited:", current_move_list)
-                    # q.append((i, path + [curr]))
-                    #visited.add(i)
-                                             # new_q.append((i, path + [i]))
-                    q.append((i, path + [curr]))             
-                    print("q after app:", q)
-            '''if not q:
-                move_list = path[:] + [curr]'''
-            self.find_path_bfs(new_grid, new_move_list, q)
+        explored = []
+        queue = [[start]]
+        print("goal:", goal)
+        while queue:
+            path = queue.pop(0)
+            node = path[-1]
+            map.truck.move_to(node[0], node[1])
+            if node not in explored:
+                possible_cord = map.truck.possible_cord()
+                for i in possible_cord:
+                    new_path = list(path)
+                    new_path.append(i)
+                    queue.append(new_path)
+                    explored.append(node)
+                    trash_around = map.truck.find_trash(type_of_trash, i[0], i[1])
+                    if trash_around != []:
+                        print(new_path)
+                        tmp = map.truck.find_trash(type_of_trash, i[0], i[1])
+                        trash_to_collect = tmp[0]
+                        map.truck.collect_trash(trash_to_collect)
+                        move_list += new_path[1:]
+                        move_list.append("collect")
+                        #move_list.append(i)
+                        self.bfs_shortest_path(map,  new_path[-1], map.truck.find_next_trash_to_visit('yellow_trash') )
 
 
-        '''trash_around = grid.truck.find_trash_around(type_of_trash)
-            if trash_around != []:
-                for i in trash_around:
-                    move_list.append('collect')
-                    tmp = grid.truck.find_trash_around(type_of_trash)
-                    trash_to_collect_on_new_grid = tmp[0]
-                    grid.truck.collect_trash(trash_to_collect_on_new_grid)
-                    
-            else:
-                possible_moves = grid.truck.possible_moves()
-                if possible_moves == []:
-                    return
-                for i in possible_moves:
-                    if not i == last_move and i not in visited:
-                        q.put(i)'''
-                       
-    
-    def check_if_is_done(self, m):
+    def check_if_is_done(self,m):
         for i in m.grid:
             for j in i:
                 if j.get_type() == type_of_trash:
