+++ main.py
@@ -24,7 +24,8 @@ for i in _map.grid:
             j.type = "yellow_trash"
 
 #move_list = DeepFirstSearch().start_dfs(_map, 'yellow_trash')
-move_list = BestFirstSearch().start_A_star(_map, 'yellow_trash')
+#move_list = BestFirstSearch().start_A_star(_map, 'yellow_trash')
+move_list = BreathFirstSearch().start_bfs(_map, 'yellow_trash')
 
 print("Końcowa lista ruchów")
 
@@ -39,9 +40,10 @@ while(True):
             tmp = truck.find_trash_around("yellow_trash")
             truck.collect_trash(tmp[0])
         else:
-            truck.make_move(move_list[j])
+            #truck.make_move(move_list[j])
+            truck.move_to(move_list[j][0], move_list[j][1])
 
         j = j+1
 
     _map.update_window()
-    _map.render_window(
+    _map.render_window()
